js folder for canvas
