+++ app/javascript/packs/canvas.js
@@ -1,0 +1,1 @@
+import '../components/canvas-main';
