+++ app/views/pages/home.html.erb
@@ -10,3 +10,4 @@
     <div class="bottom-navigation"></div>
   </div>
 </div>
+<%= javascript_include_tag 'canvas' %> <%= javascript_pack_tag 'canvas' %>
